structure component into independent files
